Drops dependency from VL.Skia.Core to SharpDX.Direct3D11
By using the native Win32 API (when running in Windows) we can leave it up to ANGLE to setup the Direct3D11 device per thread.
This also opens the option for direct composition / flip model in swap chain because only when the EGL display is created by ANGLE some internal flags are enabled.