Fixes `FileTexture` no longer working in exported apps
There was an unnecessary reference to `Microsoft.Build` which caused a `PackageReference` in the generated C# project file.
As documented in https://learn.microsoft.com/en-us/visualstudio/msbuild/find-and-use-msbuild-versions?view=vs-2022#use-microsoftbuildlocator such a reference (if it exists) must be a compile time dependency only. If the runtime assembly ends up in the output folder the load of msbuild fails (see `PackageSessionPublicHelper` in `Stride`).