The "default" of a channel is the dummy channel - fixes slightly different behavior compared to state before input API changes
Original report is https://discourse.vvvv.org/t/imgui-tabitem-is-broken-in-latest-preview/21614
Note that the TabItem has a "Visible" input which is set to FALSE, while it should be true. Even in 5.1 this leads to a weired user experience, the pin not reacting on first change and also not working when getting reset. Basically it only works because of some plumbing code in the node itself. This should get reviewed.