[CoreLib] Path no longer caches its properties. Fixes super weired file exists bugs / crashes.
The commit also changes the Children property to NOT be recursive (no idea why it was ever checked in like that).
We should further change the return type of Children and others to IEnumerable (and NOT spread), to allow lazy lookups / filtering.
Further cleanup should be done around the CopyTo methods - they return some weired funcs