[RedisModuleClientSideCaching]
Update to StackExchange.Redis 2.7.4
https://stackexchange.github.io/StackExchange.Redis/ReleaseNotes#274
Enables the complete conversion to ClientSideCaching
This is how we get rid of the Custom IsChanged implementation.

All critical methods in the RedisCommandQueue and SubscriberExtension
implement the NodeContext and IVLRuntime.Current,
so that PersistentMessage can also be promoted to VL from Observables
that run concurrently.

Add HelpPatch for "HowTo External Write to Global Channel.vl"

TODO:
Implement SerializerEnum(MsgPack, JSon, None)
in RedisClient/RedisModule,
in Binding "Default, MsgPack, JSon, None" and
in Publish and Subscribe.
This will be done when VL.MessagePack is ready.