Adds the app host to the node context
It makes sense to have access to the host from within the node context (transporting information about the environment where the node is created). That way we can use it to retrieve a logger bound to the proper host and with proper node location.

The static field `NodeContext.Default` is now obsolete and must no longer be used. It's only there for ABI compatibility.