[Skia] Fixes crash / slow rendering due to unintended device sharing between threads
Even though our code explicitly said that a new device has to be created for each thread, it somehow no longer worked. Debugging the ANLGE code showed that it would simply take the previously created device. We now no longer rely on ANGLE's device creation but instead create it on our own (thanks to CsWin32 for the easy way of doing that without pulling in megabytes of dependencies).