Speeds up Synchronizer in mutable scenarios
Adds a second internal implementation using a mutable dictionary. The mutable or immutable implementations are chosen based on the immutability flag of the patch the node is used in.
In a patch creating a few hundred global channels and tracing it for a couple of seconds the time spent in the Synchronizer class went down from 1300ms to 250ms.