[Stride] Don't use SDL For non-user patches (extensions, tooltips, etc.)
SDL assumes one single main thread. It can't deal with multiple main loops. This lead to crashes / freezing windows when Stride was used in both the user and editor patches.
Examples are VL.SpoutMonitor.HDE, VL.Audio.HDE or the tooltips of VL.Fuse on constant input values.

The Windows Forms backend of Stride doesn't suffer from those issues, and works normally with two main threads. However it has other drawbacks, like no touch input (there is a pending PR of ours) and fullscreen doesn't seem to set the window position properly.
Both of those issues are of not that big of a concern for the above usage scenarios.