One Direct3D11 device per app, graphics resources can therefor be shared nicely between Skia and Stride
While it was possible to render Skia inside of Stride, one would hit a road block when that Skia graph was accessing GPU backed resources. For example https://forum.vvvv.org/t/skia-resize-prevents-layer-to-be-rendered-in-stride/20202
In this approach we register Stride as the primary device provider. Skia will test for that, and if present, use the existing device instead of creating its own (which was happing until now).