tries to make the channel node the Owner of a public channel.
problem: Tries to clarify ownership within a cooperative system which leads to a class of new issues. (picture a patch with 2 nodes: Timeline & Channel node. Can we ensure that the order of execution is irrelevant?) My current take on it: Better both nodes act the same way in a symmetric fashion. Not one node trying to act as more important than the other.