[Stride] Simplifies GPU viewer - it was using a SkiaRenderer just to create a 512x512 texture
The initial issue we are investigating is https://forum.vvvv.org/t/gpu-rgba-tooltips-rendered-transparent-in-preview-495/24534
We saw it also happens on other machines now. Looking at the log, the SkiaRenderer inside that used Rectangle/TextureFX node was crashing.
We're not sure yet if this was the root cause for the transparent tooltip - we can't reproduce it anymore - even on unchanged old version.