[Redis] Returned `RedisClient` object refers to the node of the same name and is always assigned
Restructures code so that `RedisClient` is the main class, managing a connection and all the bindings.
The structure before was rather convoluted with lot's of back-pointers (`RedisModule` -> `RedisClientManager` -> `RedisClient` which was passed around in user land and null for async connects)

With those changes in place, the persistent bindings (operating on public channels) can also exist the moment the RedisClient node starts. Before they did not exist until a connection was available.
The tooltip on the column will say if the client is not connected.