[Stride] Fixes TextureFX nodes not cleaning up their effect instance (thanks to new refcnt)
Note however that there's still an internal pipeline state getting created via a cache which only gets cleaned up on F8.
This means that when we constantly create and destroy patches containing texture FX nodes we have a leak until we stop the app.

Internally Stride caches pipeline states, but by using the EffectInstance class a new RootSignature gets created which in turn leads to a cache miss.
Question is if the RootSignature could be used across effect instances.